complete game10

+++ GAME10/MISSILE_BULLETS.h
@@ -5,7 +5,7 @@ class MISSILE_BULLETS :public BULLETS
 public:
     MISSILE_BULLETS(class GAME10_GAME* game);
     void create();
-    void launch(const VECTOR2& pos, int speed, int lane);
+    void launch(const VECTOR2& pos, float speed, int lane);
     void update();
 };
